UPDATE: implement PattenService

+++ Presentation/ScalesDesktop/Features/Pallet/PalletModel.cs
@@ -4,7 +4,7 @@ namespace ScalesDesktop.Features.Pallet;
 
 public class PalletModel
 {
-    public int Number { get; set; } = int.MinValue;
+    public long Number { get; set; } = int.MinValue;
     public Guid Uid { get; set; } = Guid.Empty;
     public DateTime CreateDt { get; set; } = DateTime.Now;
     public IEnumerable<LabelEntity> Labels { get; set; } = [];
